Refactor fallback runner to modular command builder
